requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
DateTime==5.2
idna==3.4
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
telebot==0.0.5
urllib3==2.0.7
zope.interface==6.1
